Restructured app
Made major changes to the Calculator class in anticipation of
trigonometric functions; it now stores data in an Equation rather than
a single string. Created the Equation object which extends an array
list of strings. Added many functions to the Equation class to make the
Calculator class more readable. Bug fixes and other improvements.